// school-frontend/src/api/school.ts
export interface Course {
  course_id?: number | null;
  instructor_id: number;
  title: string;
  course_number: string;
  credits: number;
}

export interface CreateCourseRequest {
  instructor_id: number;
  title: string;
  course_number: string;
  credits: number;
}